The sizes are used by trillian client

+++ protocol.h
@@ -207,8 +207,8 @@ enum ERROR: uint16_t {
 }
 
 namespace IM {
-const int CAPABILITY_IM     = 1;
-const int CAPABILITY_TYPING = 2;
+const uint16_t CAPABILITY_IM     = 1;
+const uint16_t CAPABILITY_TYPING = 2;
 
 /* a tlv_packet msg_type */
 enum MSG_TYPE: uint16_t {
